module1/requirements.txt
pandas
numpy
scikit-learn
scipy
joblib
pyyaml
matplotlib
